Function rendy_command::families_from_device

pub unsafe fn families_from_device<B>(
    device: DeviceId, 
    queue_groups: &mut Vec<QueueGroup<B>>, 
    families: impl IntoIterator<Item = (FamilyId, usize)>, 
    queue_types: &[impl QueueFamily]
) -> Families<B> where
    B: Backend, 

Query queue families from device.

Safety

This function shouldn't be used more then once with same parameters. Raw queue handle queried from device can make returned Family usage invalid. families iterator must yeild unique family indices with queue count used during device creation. properties must contain properties retuned for queue family from physical device for each family id yielded by families.