Struct rendy_command::Family

pub struct Family<B: Backend, C = QueueType> { /* fields omitted */ }

Family of the command queues. Queues from one family can share resources and execute command buffers associated with the family. All queues of the family have same capabilities.

Methods

impl<B, C> Family<B, C> where
B: Backend,

pub fn device_id(&self) -> DeviceId

Get owned id.

pub fn assert_device_owner(&self, device: &Device<B>)

Assert specified device is owner.

pub fn instance_id(&self) -> InstanceId

Get owned id.

pub fn assert_instance_owner(&self, instance: &Instance<B>)

Assert specified instance is owner.

impl<B> Family<B, QueueType> where
B: Backend,

pub unsafe fn from_device(
queue_groups: &mut Vec<QueueGroup<B>>,
id: FamilyId,
count: usize,
family: &impl QueueFamily
) -> Self

Query queue family from device.

Safety

This function shouldn't be used more then once with the same parameters. Raw queue handle queried from device can make Family usage invalid. family must be one of the family indices used during device creation. properties must be the properties retuned for queue family from physical device.

impl<B, C> Family<B, C> where
B: Backend,

pub fn id(&self) -> FamilyId

Get id of the family.

pub fn queue(&self, index: usize) -> &Queue<B>

Get queue by index

pub fn queue_mut(&mut self, index: usize) -> &mut Queue<B>

Get queue by index

pub fn as_slice(&self) -> &[Queue<B>]

Get queues of the family.

pub fn as_slice_mut(&mut self) -> &mut [Queue<B>]

Get queues of the family.

pub fn create_pool<R>(
&self,
device: &Device<B>
) -> Result<CommandPool<B, C, R>, OutOfMemory> where
R: Reset,
C: Capability,

Create command pool associated with the family. Command buffers created from the pool could be submitted to the queues of the family.

pub fn capability(&self) -> C where
C: Capability,

Get family capability.

pub fn with_queue_type(self) -> Family<B, QueueType> where
C: Capability,

Convert capability from type-level to value-level.

pub fn with_capability<U>(self) -> Result<Family<B, U>, Self> where
C: Supports<U>,

Convert capability into type-level one.