Trait rendy_command::Submittable

pub unsafe trait Submittable<B: Backend, L = PrimaryLevel, P = OutsideRenderPass> {
    fn family(&self) -> FamilyId;

    unsafe fn raw<'a>(self) -> &'a B::CommandBuffer;
}

Submittable object. Values that implement this trait can be submitted to the queues or executed as part of primary buffers (in case of Submittable<B, SecondaryLevel>).

Required methods

fn family(&self) -> FamilyId

Get family that this submittable is belong to.

unsafe fn raw<'a>(self) -> &'a B::CommandBuffer

Get raw command buffer. This function is intended for submitting command buffer into raw queue.

Safety

This function returns unbound reference to the raw command buffer. The actual lifetime of the command buffer is tied to the original CommandBuffer wrapper. CommandBuffer must not destroy raw command buffer or give access to it before submitted command is complete so using this funcion to submit command buffer into queue must be valid.

Loading content...